Move KeyboardAvoidingView where it belongs
