+++ src/pages/NewGroupPage.js
@@ -17,7 +17,7 @@ import FullScreenLoadingIndicator from '../components/FullscreenLoadingIndicator
 import withLocalize, {withLocalizePropTypes} from '../components/withLocalize';
 import compose from '../libs/compose';
 import Button from '../components/Button';
-import KeyboardAvoidingView from '../libs/KeyboardAvoidingView';
+import KeyboardAvoidingView from '../components/KeyboardAvoidingView';
 import FixedFooter from '../components/FixedFooter';
 
 const personalDetailsPropTypes = PropTypes.shape({
