+++ src/pages/settings/Profile/ProfilePage.js
@@ -30,7 +30,7 @@ import Picker from '../../../components/Picker';
 import withLocalize, {withLocalizePropTypes} from '../../../components/withLocalize';
 import compose from '../../../libs/compose';
 import Button from '../../../components/Button';
-import KeyboardAvoidingView from '../../../libs/KeyboardAvoidingView';
+import KeyboardAvoidingView from '../../../components/KeyboardAvoidingView';
 import FixedFooter from '../../../components/FixedFooter';
 import Growl from '../../../libs/Growl';
